Add flag to curl invocation to follow redirects
Because GitHub has evidently changed the location of the referenced files and is redirecting to them with a 301 Moved Permanently, the command lines as-is will not work today. Adding the -L flag to follow redirects seems to do the trick, though you may want to consider also replacing the URLs with their new canonical values.

Thanks for the great tool!